backend/bigquery-sync/requirements.txt
google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
pandas==2.1.4
functions-framework==3.5.0
pyarrow==14.0.1